day9/pyproject.toml
[project]
name = "mongodb-crud-cli"
version = "0.1.0"
description = "A command-line interface CRUD tool for MongoDB."
authors = [
    { name = "Paramswari", email = "parameswaribala@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [       
    "pymongo>=4.0.0",
    "python-dotenv>=0.19.0",
    "certifi>=2024.6.0"
   
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.0.0"
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py313"]
[tool.isort]
profile = "black"
